Add ipv6 dns configs to cloud
