+++ lte/gateway/python/magma/mobilityd/main.py
@@ -10,12 +10,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+import ipaddress
 import logging
+from typing import Optional
 
 from magma.common.redis.client import get_default_client
 from magma.common.service import MagmaService
 from magma.common.service_registry import ServiceRegistry
 from magma.mobilityd.ip_address_man import IPAddressManager
+from magma.mobilityd.ip_allocator_base import OverlappedIPBlocksError
 from magma.mobilityd.ip_allocator_dhcp import IPAllocatorDHCP
 from magma.mobilityd.ip_allocator_pool import IpAllocatorPool
 from magma.mobilityd.ip_allocator_static import IPAllocatorStaticWrapper
@@ -61,6 +64,27 @@ def _get_ipv4_allocator(store: MobilityStore, allocator_type: int,
     return ip_allocator
 
 
+def _get_ip_block(ip_block_str: str) -> Optional[ipaddress.ip_network]:
+    """ Convert string into ipaddress.ip_network. Support both IPv4 or IPv6
+    addresses.
+
+        Args:
+            ip_block_str(string): network address, e.g. "192.168.0.0/24".
+
+        Returns:
+            ip_block(ipaddress.ip_network)
+    """
+    if not ip_block_str:
+        logging.error("Empty IP block")
+        return None
+    try:
+        ip_block = ipaddress.ip_network(ip_block_str)
+    except ValueError:
+        logging.error("Invalid IP block format: %s", ip_block_str)
+        return None
+    return ip_block
+
+
 def main():
     """ main() for MobilityD """
     service = MagmaService('mobilityd', mconfigs_pb2.MobilityD())
@@ -99,10 +123,26 @@ def main():
     ip_address_man = IPAddressManager(ipv4_allocator, ipv6_allocator, store)
 
     # Add all servicers to the server
-    mobility_service_servicer = MobilityServiceRpcServicer(ip_address_man,
-                                                           mconfig.ip_block,
-                                                           mconfig.ipv6_block)
+    mobility_service_servicer = MobilityServiceRpcServicer(ip_address_man)
     mobility_service_servicer.add_to_server(service.rpc_server)
+
+    # Load IPv4 and IPv6 blocks from the configurable mconfig file
+    # No dynamic reloading support for now, assume restart after updates
+    logging.info('Adding IPv4 block')
+    ipv4_block = _get_ip_block(mconfig.ip_block)
+    if ipv4_block is not None:
+        try:
+            mobility_service_servicer.add_ip_block(ipv4_block)
+        except OverlappedIPBlocksError:
+            logging.warning("Overlapped IPv4 block: %s", ipv4_block)
+
+    logging.info('Adding IPv6 block')
+    ipv6_block = _get_ip_block(mconfig.ipv6_block)
+    if ipv6_block is not None:
+        try:
+            mobility_service_servicer.add_ip_block(ipv6_block)
+        except OverlappedIPBlocksError:
+            logging.warning("Overlapped IPv6 block: %s", ipv6_block)
 
     # Run the service loop
     service.run()
